+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/PneuResource.java
@@ -113,27 +113,4 @@ public final class PneuResource {
         service.marcarFotoComoSincronizada(codPneu, urlFotoPneu);
         return Response.ok("Foto marcada como sincronizada com sucesso");
     }
-
-    /**
-     * @deprecated at 2018-08-22. Utilize {@link #marcarFotoComoSincronizada(Long, String)}.
-     * Este método ainda é mantido para permitir que apps antigos sincronizem suas fotos.
-     * Como este resource foi liberado apenas para versões do app > 57, nós adicionamos o
-     * {@link AppVersionCodeHandler} neste método para permitir que apenas ele tenha um tratamento diferente, permitindo
-     * que a sincronia das fotos aconteça para aplicativos antigos (version code > 55). Isso funciona pois o
-     * {@link AppVersionCodeHandler} prioriza anotações a nível de método.
-     */
-    @PUT
-    @Secured
-    @Path("/unidades/{codUnidade}/{codPneu}/fotos-cadastro/sincronizada")
-    @AppVersionCodeHandler(
-            implementation = DefaultAppVersionCodeHandler.class,
-            targetVersionCode = 55,
-            versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
-            actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
-    public Response DEPRECATED_MARCAR_FOTO_COMO_SINCRONIZADA(@PathParam("codUnidade") @Required final Long codUnidade,
-                                                             @PathParam("codPneu") @Required final Long codPneu,
-                                                             @QueryParam("urlFotoPneu") @Required final String urlFotoPneu) {
-        service.marcarFotoComoSincronizada(codPneu, urlFotoPneu);
-        return Response.ok("Foto marcada como sincronizada com sucesso");
-    }
 }